# zentra/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateEmployeeRequest:
    description: Create employee request model
    properties:
      DivisionID:
        example: 1
        type: integer
      email:
        example: alice.johnson@email.com
        type: string
      name:
        example: Alice Johnson
        type: string
      phone:
        example: 123-456-7890
        type: string
    required:
    - DivisionID
    - email
    - name
    type: object
...
  handlers.DivisionInfo:
    properties:
      description:
        example: Administration department
        type: string
      id:
        example: 1
        type: integer
      name:
        example: ADMIN
        type: string
    type: object
...
  handlers.EmployeeResponse:
    description: Employee response model
    properties:
      created_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      created_by:
        example: admin
        type: string
      division:
        $ref: '#/definitions/handlers.DivisionInfo'
      division_id:
        example: 1
        type: integer
      email:
        example: alice.johnson@email.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Alice Johnson
        type: string
      phone:
        example: 123-456-7890
        type: string
      tenant_id:
        example: 1
        type: integer
      updated_at:
        example: "2024-03-24T21:41:49Z"
        type: string
      updated_by:
        example: admin
        type: string
    type: object
  handlers.ErrorResponse:
    description: Standard error response model
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.SuccessResponse:
    description: Standard success response model
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.UpdateEmployeeRequest:
    description: Update employee request model
    properties:
      DivisionID:
        example: 1
        type: integer
      email:
        example: alice.johnson@email.com
        type: string
      name:
        example: Alice Johnson
        type: string
      phone:
        example: 123-456-7890
        type: string
    required:
    - DivisionID
    - email
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Vomo backend API documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Vomo API
  version: "1.0"
paths:
...
  /employees:
    get:
      description: Get all employees
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.EmployeeResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all employees
      tags:
      - Employee
    post:
      consumes:
      - application/json
      description: Create a new employee with the provided details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Employee Data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateEmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.EmployeeResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new employee
      tags:
      - Employee
  /employees/{id}:
    delete:
      description: Delete an existing employee
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an employee
      tags:
      - Employee
    get:
      description: Get employee details by ID
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EmployeeResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an employee by ID
      tags:
      - Employee
    put:
      consumes:
      - application/json
      description: Update an existing employee with new details
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Employee ID
        in: path
        name: id
        required: true
        type: integer
      - description: Employee Data
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.EmployeeResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Employee not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an employee
      tags:
      - Employee
  /employees/by-division:
    get:
      description: Get all employees for a specific division
      parameters:
      - description: Tenant ID
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Division ID
        in: query
        name: division_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.EmployeeResponse'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get employees by division
      tags:
      - Employee
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"